Added `ser/serializer` module
